AFFINITY-26 Add a faster JNI timer and performance tune a number of key benchmarks.

+++ affinity/src/main/java/net/openhft/clock/IClock.java
@@ -26,5 +26,5 @@ public interface IClock {
 
     long toNanos(long ticks);
 
-    double toMicros(long ticks);
+    double toMicros(double ticks);
 }
